refactor(verifier): modularize bot.py into internal package\n\n- Add verifier_bot/ with logging_utils, coc_api, approvals\n- Keep bot.py as thin facade; preserve public API for tests\n- Update Dockerfile to copy new package\n- Robustify giveawaybot schedule_check month-end calc\n- Packaging: configure setuptools to include verifier_bot and top-level modules\n\nAll tests pass (188) with >=80% coverage.
