+++ Dockerfile
@@ -10,4 +10,5 @@ WORKDIR /app
 COPY --from=dependencies /usr/local/lib/python3.12/site-packages /usr/local/lib/python3.12/site-packages
 COPY --from=dependencies /usr/local/bin /usr/local/bin
 COPY bot.py .
+COPY verifier_bot ./verifier_bot
 CMD ["python", "bot.py"]
