+++ pyproject.toml
@@ -15,6 +15,18 @@ dependencies = [
     "boto3>=1.34",
 ]
 
+[tool.setuptools]
+# Treat top-level .py files as modules and only package our internal library.
+py-modules = [
+    "bot",
+    "giveawaybot",
+    "giveaway_fairness",
+    "fairness_simulation",
+]
+
+[tool.setuptools.packages.find]
+include = ["verifier_bot*"]
+
 [project.optional-dependencies]
 dev = [
     "pytest>=7.0.0",
