+++ verifier_bot/__init__.py
@@ -1,0 +1,5 @@
+"""Internal modules for the verifier bot.
+
+This package holds implementation details split out of bot.py to keep the
+entrypoint thin while preserving the public API expected by tests.
+"""
